Adds github actions for mainline and for testing
Testing includes some best-guess tests for syntax, checking the
validity of the inventory, etc.

Room for improvement is the disconnect between github actions' way of
getting the Ansible inventory submodule in the right spot (discrete
cloning of repo to specified directory) versus the way developers might
do it (git submodule). Unify it with submodules or do it with a discrete
clone, but choose one and do it consistently! ^_^;;;